+++ packages/kit/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @logos-ui/kit
 
+## 2.0.8
+
+### Patch Changes
+
+- Updated dependencies [2b64cfa]
+  - @logos-ui/fetch@3.0.0
+
 ## 2.0.7
 
 ### Patch Changes